Solution for day1 in rust and python
Using very similar constructs, did the python variant because I was stuck
in rust for a little while.

No functions as I didn't look into how to pass the iterator or list around
to a function, though I think now that I have the solution it ought to be
quite easy.