Solution for day2, rust and python
The python solution only does part2 as that's where I struggled with my
rust-foo. The rust code for part2 (and maybe part1) is likely extremely
in-efficient in terms of memory-allocations - looking at the
r.chars().nth() in the findmismatch helper, but it's too late now to
care for memory efficiency.