Add a python version
This is easier to experiment different ways of solving the issue which
I can then port to rust to learn a bit more of the language. Already had
to make one adjustment: Change from recursion to an iterative approach for
accumulating the overlapped points as python's recursion limit is hit with
the input data.