Fix off-by-one error in output
There was no logic error at all, except that with the way I used the
summed-area algorithm I ended up returning the coordinate that is to the
top and left of the area for which I calculated the summed power level.
This is visible in the wikipedia article's graphics, the 'D' point (the
topleft edge of the area) is always outside of the actual coordinates being
summed up.

So fix that by adding to the coordinate pair when returning it.