Solve part2 of day12 in constant time
After getting some hints towards this, the trick here is that the
pattern generated by the plants stabilizes after a set of initial
iterations and the sum ends up growing by a constant factor from that
point onwards.

The point where the stabilization happens is within a range that allows
to run the algorithm in brute force, I spotted the point by inspecting
the pattern and then substracting a few sums to see where the actual point
is. The fixed numbers for part two are based on this observation.

Since my generations start at 0 the 50-billionth one is 50-billion minus 1,
just to clarify where the -1 is coming from. Can be verified by checking
the output for part1 where the right answer is in generation number 19.