Day18
For part2 found a loop visually, then calculated startpoint and based on
that put up a formula that'll figure out when to stop calculating the
resource value since it repeats

Had an off-by-one in the targetminute-calculation, likely due to i
starting at minute '0', but I didn't further think of the actual reasons.
Just happened to notice that the previous value is larger and that's what
the website told me about my wrong value.