Implement part2 by speeding up the inner loop
After some help from friends and a bit of cheating on reddit I finally
was able to see that 'inner' loop and reverse-engineer it into a higher
level language such that I can understand what it tries to do. Then finding
a replacement was relatively simple. Ignoring the instructions of the
inner loop and letting python's modulo handle that is fast enough to finish
calculation of the program in about 30 seconds.

I'm also adding the 'initialization' block and the values it produced as
well as the loop and my pseudocode I wrote while reverse-engineering the
'assembler'